crash fix and clay brush strength fix
Completely removed thread safe path in calc_area_normal that was causing crash when used with multires.  Only half-way making the necessary changes to disable it is what created the unstable code.

Made clay brush strength dependent on the projected radius of the brush so that it has the same apparent effect no matter what scale the user zooms.