Render Branch: sss optimization, to share irradiance cache between materials in preprocessing pass. This required fairly major changes, so that each tile renders all sss materials, instead of each material rendering all tiles.
I tried further to do a single rasterization for all materials, but it changed
the results too much, for example eyes would be missing shadow at the sides
because occluded faces were included in the sss pass then and not contributing
any shadow. Unfortunate because it helped performance quite a bit as exactly in
these cases the irradiance caching needs to take more samples.