== Layer Modifier Brushes ==
For all tools that it makes sense for, the 'Use Layer' modifier allows you to set a limit on how far a vertex moves which creates a layering effect.

Also, added a slider to control how thick a layer is.

Also, Fix #23192, where layers were not working with symmetry.

After testing more will apply to trunk.