MapUV Node - Tile
This node requires an EWA filter to read the final value.
In COM_MemoryBuffer.cpp I copied the functions from texture.c
The reason for copying and not re-using from a header is to allow
for further refactoring. Basically the codes should be independent.

The EWA happens automatically when  you call the read function with two extra arguments
(dx and dy, calculated from a differential filter pass before in the operation).

Good news, the result works better than trunk. In trunk the edges of the image get black,
here it works all fine.