Masks tracking operators
Added an operators which tracks the whole masks
in clip editor. Namely, this operators are tracking
selected splines of active spline layer.

This uses the same exact region tracker as motion
tracker used for ages, just marker is created from
mask spline. Pattern area will match the size of
spline's bounding box (without feather taken into
account!). Search area is currently hardcoded to
1.5x times bigger than marker's pattern area.

Settings for the tracked are used from default
tracker settings (which are in T-panel). So for
this added special panel to mask mode which
contains default tracker settings.

Matching will be ignored and Previous Frame
matching will be used.

Speed is also currently hardcoded to realtime.

This new tracker operators are using the same
shortcuts as regular tracker.

TODO:
  - Feather support.
  - Better deal with search area size.
  - Think of tracking speed control.

This is an experimental feature, might crash or
do not behave as you expect it to behave.