Fix old undo issue for projective texture paint
Undoing when faces accessed the same region in uv space could create
artifacts. The issue here is that buckets get their original colour
initialized with the colour that exists in the image at the point of
initialization. If two faces on different buckets accessed the same
region, then one bucket would get the modified colour as the original
colour. The undo push code simply checked if the final colour was
different than the original colour and as a result it could copy
modified colours in the undo stack.

Now simply push the undo stack when initializing the buckets. This
ensures that the stack contains really original colours. Of course there
is some overhead.

Also added mask code creation as part of the undo tile creation. The
accumulated mask should be in canvas space pixels, not on screen space
bucket pixels. I will next hook up the projection pixels to reference
this mask. This will hopefully solve cases like painting on mirrored
meshes or meshes that the same canvas region is accessed by many faces.