Anchored brushes for texture paint.
To support this I added size support to the RNA stroke so the stroke
size can be calculated in the stroke system and passed on to paint
systems. Also, the output mouse coordinate is directly modified now and
the size is tweaked accordingly. Looks sculpt does not directly use this
for anchored strokes as far as I have seen, but accidents could happen
there so I hope there will be some artist testing.

Also fixed error in texture coordinates for jittered strokes.