Polyline becomes a separate stroke mode. Rationale for this is that artists like the quick drag and release for quick lines and requiring the user to keep the shift key pressed for new line segments would be cumbersome and inconsistent because it would mean that first line additionis done on release and subsequent get added on key press.
So, lines now use a drag-release paradigm and polylines use regular
clicking. To confirm, use space or enter like on any operator. Polyline
bit was used for fixed textures on earlier blender versions, meaning
that I'll have to make sure that flag is not used anymore. I will do a
commit for this on trunk.