Changes to the fill/gradient tools:
Fill is now a paint tool, this is done to get correct cursor drawing for
gradient fill. It also allows to unify the gradient and fill tools to
one and use the appropriate depending on whether gradient is checked or
not.

Gradient fill and strokes are still inactive in this commmit, but I
added another goodie, drag'n'drop of colors in the image editor main
window while paint mode is active will do bucket filling.