Change the way projective texture painting sets up the paint layers.
Projective texture painting now includes a material list in the
projection painting options in the toolbar. By selecting a material in
this menu, the user can see the avalailable texture layers that can be
used for painting (Those that are of image type and have UV mapping
currently, as expalined by a label below) in a separate texture list.

http://www.pasteall.org/pic/58315

By selecting textures in that list, the user can now immediately select
a texture slot of the material to paint on. This means that painting now
has immediate effect on the output of rendering :)

Now time to gather artist feedback.

TODO:
* Improve behaviour of projection painting so that it uses the UV layer
of the texture slot, if it exists, instead of the active UV layer.

* Find a way to cache the active paint image of the material. There are
a few places this can get invalidated. I'd like a central way to handle
this but probably a case by case solution will have to be made

* Add more user convenience functions: Operators to add material layers
for painting, automatically adding layers/materials if they do not
exist, maybe automatic UV unwrapping? With appropriate warning that for
optimal results a manual unwrap is needed.