Add layer operator in projective texture paint layer panel.
This commit adds a new mtex slot of type image, using UV mapping on
active material. It also allocates a new image for it. Size is fixed
currently (1024x1024) but this can change later.
This code can surely be useful to enforce a working texture paint system
as soon as user enters.I hope resource management is done right here, it
looks like it works (tm).