Removed remaining pointcache->mem_cache access in particles. Instead the particle system should now use its own local mem_cache list for storing cache data in memory. The actual PointCache struct does not make this distinction any more.
Eventually the particle system should have clean methods for reading
cache data into other structures for editing and displaying cache data,
but this can not be cleaned up easily without major refactoring. For now
the important thing is to keep psys away from the PointCache internals
so we can implement a nice API for it without psys interfering.