Cycles Volume: materials without a surface shader now act transparent when there is a volume shader.
This is the behavior you would expect I think and makes setup easier. More
optimizations are possible in this case to avoid surface shading altogether, but
it's probably not so noticeable altogether as this will early out and not get to
any expensive computations as far as I know.

Note that these still count as transparent bounces even if there is no BSDF. Not
entirely happy with that, but there should be some way to limit the depth?