Marked the PTC_STATE_SIMULATION_VALID flag as deprecated. This is only used by the bake operator, but keeping track of the sync state of cache data vs. simulation settings needs to be radically simplified.
The API should provide a way of detecting whether a valid simulation
state can be obtained for a certain frame.

  /* check if a frame can be read from the cache (no data is modified) */
  PTCReadSampleResult PTC_test_sample(struct PTCReader *reader, float frame);
  /* try to obtain a cache sample and write to DNA data */
  PTCReadSampleResult PTC_read_sample(struct PTCReader *reader, float frame);

Beyond that the cache should be oblivious of the simulation state. When
settings or dependencies are changed which invalidate the simulation
cache state this should invalidate all cache samples (or possible a
certain frame range), but the cache itself should not need to keep track
of the simulation as a whole.