Removed the BAKED flag from point cache state. The cached now has a new flag LOCK_SETTINGS instead, which should indicate more clearly that it is about preventing cache modification by disabling user settings changes.
The cache should not have to care about this setting in the first place.
It becomes increasingly difficult to control changed simulation settings
anyway (scripting!), and the cache should not even try to exert control
over sims this way. Eventually this flag should be a feature of
simulations and modifiers, using it as a 3rd option beside
enabled/disabled and using only existing cached data.