Rename evaluation_context to eval_ctx
Only affects on variable name, structure name stays the same.
Addresses code review from Campbell to make variable names
which are being passed all over be short.

Not entirely happy with this, but well ok. It's still verbose
enough and naming kind of consistent to what Python uses.

Discussed with Brecht and he's also fine with such a change.

Hopefully this is last change to be done before merge :)