Reusable curves for the paint systems:
This commit includes the following:

* Per brush curve setting that allows setting a different curve per brush.
Those curves stay in memory and can be edited and reused accordingly.
To make sure that a curve is saved in a file users need to add a fake user,
similar to other datablocks.

* Interface while using curves is now non blocking. Users can pan, rotate,
and zoom the view while able to select and edit curve points as previously
in the GSOC branch.
Code internally uses the transform system. This needs some tweaking still
but it should be similar to the previous GSOC system. This may change
in the future to be more compatible with other curve systems
(ie enable choosing handle type etc.)