Support proper display of different UV layers.
It uses a little DAG hack to force mesh update when tweaking the material.
This is not really necessary, and it would be good if there was a way to
only force recreation of uv draw buffers.

Missing in this commit is support for non-VBO drawing since there's a patch
waiting for trunk that will make that code obsolete.

Subsurf will always work because it uses immediate mode drawing.