Splitting off main construction logic from depsgraph_build.cpp.
The original source file now only contains general functions for the
builder classes. The actual logic for building the nodes/relations from
a scene is in depsgraph_build_nodes/depsgraph_build_relations now. This
helps navigation and keeps the code in those files concerned only with
the actual inspection of DNA data.