Removed unused method declarations for component/operation finding directly from the graph.
The graph should rather not require such deep lookup wrappers. The use
cases for such functionality are all in the build procedure, so it makes
sense to keep the graph as a simple container with raw access only to
the root node types and id nodes. Builder classes can then provide
wrappers for detailed lookup in the respective subnodes.