Removed the `remove_from_graph` callbacks from DepsNode and the generic `remove_node` function in Depsgraph.
These were currently unused anyway. In future for removing a node the
respective remove_xxx methods in the actual container should be called.
This is more consistent with the hierarchical data structure and the
actual ownership of nodes. Might add convenience methods for this sort
of thing later, but the raw Depsgraph node container is cleaner this way.