Basic support for modifiers in the new depsgraph.
Modifiers are represented as a new geometry operation. Like with the
current modifier stacks, a typeinfo callback can be used to define
additional dependencies of a modifier on scene or object components
(currently implemented only for the shrinkwrap modifier for testing).

When this function is defined for a modifier, it gets an abstract
"DepsNodeHandle" pointer, which can be passed back to the depsgraph API
for creating the actual relations internally with minimal effort for the
modifier type and providing opportunities for internal error handling.