Pie menus beta code.
What is included here:

* Some initial code that lays out and displays items of a pie menu
 in a circular manner around the mouse. Collision detection is
 still not as expected for a pie menu.

* Code to spawn a pie menu for enums and operator enums (a bogus pie
menu for object mode is included, not yet set to any specific key).
For testing, the sculpt mode stroke mode enum is bound as a pie menu
to A key.

* Support to spawn pie menus from python.

Design could change radically, but best have the code here where all can
checkout and see the state of the code.