Change the clone tool to use the new slot system.
When clone from layer is active, a layer selection will be available in
the the layer panel and the tool will sample from the selected clone
layer.

This is not ideal UI wise, the user has to change between tabs and
panels to set the option - Maybe a good reason to rethink about setting
the layer panel to the right.

This makes the mtface_clone_layer functions and data obsolete for
texture painting.