Support edge evaluation for triangles and ngons
Still gives issues for NGons and also it's something
strange going on when applying modifier on the triangle.

Triangles seems to be looking all fine for until the
modifier is applied on the mesh, but ngons still needs
some investigation.

Also, just understood that old SS code did something
special with triangles, it didn't handle triangles in
the same way as NGons. The thing is, treating triangles
as ngons makes them looking as if some crease is being
applied on the corners.

Would need to write some special case for this in
opensubdiv case.