Stencil:
Rework projective paint stencil to use an additional image field. This
again ommits the need to go and assign an image to the UV layer in the
image editor.

Also added an extra paint mask tool. This is not yet hooked up to do
anything, but will eventually paint exclusively on the stencil image
only.