* Add custom color for new layers.
This is not ideal because of the internal representation of byte images,
which is straight alpha. This means that on the display we will get an
opaque color if we try painting on a transparent image. To bypass this I
have enforced alpha to be 1 in that case. If we are to support this we
should upload byte images as premultiplied meaning we take a performance
hit...hmmmm...no.

* Add dependency update when painting on image editor. Should handle
update of layer icons nicely