Fallback simple subdivisions to legacy code
For until OpenSubdiv case is capable of handling simple subsurf
let's fallback to legacy implementation rather than doing CC
subdivisions.

Not really sure we'll be able to have a single implementation
for bot of the cases, but bet simple subsurf might be re-done
real easy once we clean up the CCG structures for the osd.