Moved relations from generic DepsNode base class into OperationDepsNode.
Rationale here is that only relations between operations matter for the
final evaluation. Keeping this specific makes it easier to avoid
implicit assumptions and confusing upcasts in the evaluation code.

Relations between "outer" nodes (ID nodes and components) will be
translated directly into relations of their inner operation nodes (this
will require fully built operations lists at the time relations are
constructed). Higher level relations may be used for debugging purposes,
but they are not required for evaluation and should be handled in a
separate structure.