Clean up DepsNode flag fields.
The OperationDepsNode.flag field would shadow the DepsNode.flag field.
These flags are only used for eval tagging atm, so it's safe to move
these into the OperationDepsNode class.

If necessary, similar flags for "directly modified" etc. could be added
to ID nodes or components too, but they can and should be conceptually
separate and use a non-conflicting flag in the respective subclasses.