Skip DM face initialization when GPU backend is possible to use
Currently only last SS modifier in the stack is capable of using
GPU for tessellation. If there are any modifiers on top of SS then
CPU is gonna to be used (even if the modifier is disabled).

There's still uite reasonable amount of overhead happening, but
that's task for further design changes. For now let's make all
the bits working and when it's clear how it all communicates
re-design drawObject, derivedmesh and probably other areas.