Removed the build_operations methods and OperationBuilder class again.
The few places where this has been used (pose and bone components) have
been integrated into the regular build functions.

Eventually we may want to do dispatching like this again, but for the
time being it's easier to keep these things in a central place until it
is clear how the build procedure is best organized.