Moved Object->shapeflag to Key->pin
Since we aren't going to use the shape key modifier for editmode blending,
Apply Shape Keys in Editmode is now obsolete too, and shapeflag just says
is the key is pinned or not.

Why to do this? In editmode, we want to be able to switch from Blending
shapekeys to pinned shapekeys and this is done in key RNA update handler,
so we want it to be called when pinning changes.

The RNA property was removed from the Object and added to the Key, all
code that referenced Object->shapeflag now references Key->pin.

The UI was updated to accomodate for that.

Unrelated change which I couldn't stage seperately is that there is now
support for temporary mix values from RNA and UI. The user can select
which one to use in Editmode. The value in the UI will automatically
reflect the editmix or the animation mix, transparent for the artist.