Can mix shapekeys in editmode now
The user can now edit a shape key on top of a mix.

The difference to the master is that the 'Apply Shape Keys in Editmode' is gone completely. You can either pin the shapekey or edit on top of a mix: http://screenshot.su/img/a1/70/dc/a170dc61123ee1648e8a6dc3394b3812.jpg, and that is all.

The user can select the values that he wants to use in the mix: the animation-driven values (KeyBlock->curval) or a temporary mix value visible only in editmode.

Implementation-wise this is sort of an emulation of the shape keys modifier behavior. On editmode enter, the mesh keyblocks get evaluated and that is put in the edit coordinates. On shapekey switch or editmode exit the underlying mix gets subtracted from the editmesh cos, the resulting isolated shape coordinates gets adjusted for value (we store data with 1.0 values only!) and than processed as a normal (pinned) shape key update. Then the mesh gets re-evaluated, etc.

There is some optimization potential which will be explored!

Undo might be broken at the moment, this is the next thing to do.