Utility class DepsgraphOperation to bind the "context" arguments of an operation to the evaluation function.
This can replace the simple callback function and avoid anonymous void*
context arguments, which would need to be casted explicitly. Further
the context arguments and the evaluation function can be combined in a
single struct.

Eventually it would be more elegant to have true function binding, but
this has only been added to the C++ standard in C++11. We could backport
boost implementation, but until then this is a usable compromise.