Fixes and workarounds for build problems: Windows 8.1, CMake 2.8.12.2, MSVC 2013
Cmake bug(?) workaround for zlib and png:
FindPNG.cmake calls find_package(zlib) and for some reason that redundant call
causes both zlib and png to not be found properly.
(Or, at least they complain, as this does not seem to cause a build error later)
Since zlib will be found by find_package(png),
it should be OK to not call find_package(zlib) explicitly.

Setting JPEG_NAME seems to be required by any version of MSVC, not just 9,
so changed check from MSVC09 to just MSVC

The proper Boost static libraries are not found unless the following variables are set:
set(Boost_USE_STATIC_RUNTIME ON) # prefix lib
set(Boost_USE_MULTITHREADED  ON) # suffix -mt
set(Boost_USE_STATIC_LIBS    ON) # suffix -s

Note: The debug versions of the opencollada libraries are missing from windows_v12,
but this cannot be worked around (except to disable opencollada for debug builds)
because the value of _ITERATOR_DEBUG_LEVEL is different between opencollada and the
rest of Blender, so cannot just use the release versions with a debug build.
The fix would be to add the _d versions to the library directory.

Set /SAFESEH:NO to fix linker error related to libjpeg
(not sure why libjpeg is not compatible with this flag)
Also quiets warning about ignoring  /EDITANDCONTINUE due to having /SAFESEH enabled

Some typos in buildinfo.cmake fixed.

In buildinfo.cmake, trying to determine if a branch is modified when it
has no upstream branch kills the build, so added ERROR_QUIET where upstream
check was done.