Use updated version of CUEW
This wrangler is now automatically generated from
cuda.h and cudaGL.h which is really handy from the
maintenance point of view: with new CUDA toolkit
release it should be rather simple to update the
wrangler.

Currently it's generated from CUDA toolkit 5 since
it's the version which comes in Debian Wheezy.

Also bundling the generator scripts, so others
can tweak/improve it.

Works fine on Linux, didn't verify it on Windows.