Solve crash when doing viewport render
The issue was caused by the render pipeline creating
DerivedMesh for purposes of it's export to the render
database. In this case SubSurf modifier shouldn't be
applied on GPU because render pipeline would want to
access hires data from the CPU side.

Added a flag to modifier apply bitmask which indicates
that the result could be stored in the GPU side. The
flag is called MOD_APPLY_ALLOW_GPU. If this flag is
set modifiers are allowed to keep their result on GPU.

We still need to implement CPU-side hi-res mesh unpack
even if it was originally evaluated on GPU, but the
current change is useful anyway because it'll avoid
possible overhead of copying data back-n-forth.