Wrapper for std::bind, which will allow us to catch operation arguments and replace them by their respective result data types.
This is just a proof of concept, it will be used later to construct a
intermediate callback, which then performs the actual binding at
scheduling time, when all the input data of an operation is actually
available.

It uses a nifty recursion trick based on the tuple index and std::get
to apply a tuple as an argument pack in std::bind. The
bind_operation_tuple function appends tuple elements to the ArgsTail
pack one by one until N==0, at which point a template specialization is
used to issue the final std::bind call.