Simplified operation building: Use ID pointers and component identifiers directly instead of passing around nodes.
This is to some degree a revert of previous changes to add such nodes
explicitly. However, it only applies to ID and component nodes, which
are really nothing more than categories (sorting operations by ID and
components).

The only exception now are the node tree build functions, which still
take a "owner_node" and need to be looked at.