Replace region tracker with autotracker in Blender
The title actually tells it all, this commit switches
Blender to use new autotrack API from Libmv.

From the user point of view it means that prediction
model is now used when tracking which gives really
nice results.

All the other changes are not really visible for users,
those are just frame accessors, caches and so for the
new API.