(Showing My Work) Expanded GPU Module to Support Modern OpenGL
I am updating the GPU module so that it contains code needed for Blender to support a programmable shader based pipeline.

In an effort to make sure things always compile I've excluded most of the changes from 2013 that were made to gpu_codegen, gpu_draw, gpu_extensions, and gpu_material.

This new functionality is not used anywhere in Blender yet.  That will come gradually.

This is a bit rough because I wanted something in the repository before evaluations.  I wanted to break this into smaller more manageable pieces for code review, but that would take too much time.  However, this might still work for review purposes.