Pie Menus:
Add code to force a certain pie menu to be in hold style interaction
mode always. This is useful for tap/hold operator/pie menus

Move operator enum operator in C, this allows us to fire it up
internally without resorting to operators.

Deleted python code for this since it's not used and it's really easy to
code. May do if scripters want though.