Implement smooth shading for OpenSubdiv
This is rather an experiment commit which places all
the needed bits for the smooth shading in viewport.

Approach is similar to the simpleCpu example application,
which computes lo-res normal on CPU and passes them
to OpenSubdiv. Then the shader then have access to
vertex normals and passes them to the fragment shader.

At this point all the shading is smooth currently,
flat shading will be landed when all the material from
Blender are taken into account.