OpenSubdiv: Make drawing even closer to legacy viewport drawing
This actually disables using any GLSL for drawing quads in favor
of using default tessellation and shading.

This is achieved by enabling client states for vertices and
normals and this have the following benefits:

- Makes code closer to what's happening in GPU at this moment,
  hopefully this means we can use GPU_material and rest of the
  routines from this area for shading/texturing the model.

- For now it allows to have smooth and flat shading, currently
  this flag is used from the first face, need to fill in face
  flags accordingly to the original mesh.

This code is under OPENSUBDIV_LEGACY_DRAW define which makes
it rather easy to distinguish and get rid of when we'll have
viewport rrefactor in place and when it'll be more clear how
OpenSubdiv+shading/texturing would fit in there.