OpenSubdiv: Disable normals hack for legacy drawing
Such a mix of GLSL and legacy drawing was rather a huge PITA
to experiment with the new functionality. Since this only
gives wrong shading with flat shading (which isn't so common
and only really visible with simple subdivisions which are
to be redone anyway) let's do a clear separation for now.

We might end up bumping GLSL requirement for GPU tessellation
because of the some specific of how UV attributes are handled
by the OpenSubdiv. This is to be evaluated later.