OpenSubdiv: Support multi-textured faces in a solid view
This means that if faces have different materials assigned to them
they'll be textured accordingly to a diffuse settings when viewing
them in solid shading mode.

It uses approach based on the pertitioned mesh from OpenSubdiv and
currently it creates partition for every single coarse face. This
might backfire in performance for a bit because it increases number
of glDrawElements() calls. But likely partitioning will stay like
this (to avoid re-partitioning when changing material settings)
and we'll just do some conjunction of partitions when actually
drawing faces into the viewport.

Currently it's like 20% speed loss, don't get too much disappointed
on this, it's to be addressed. Just a bit later. For now 1st prio is
to have all the shading/texturing bits in place.